AudioRenderer Balance bug fix
Audio balance code seems to have a bug: using the Balance slider in WPF Sample worked well when going to Right speaker, but failed to lower the right volume when going to the Left side.
I believe it is a copy-paste error in a single line of code. This modification makes the sample project and my custom projects behave as expected with the limited number of videos files I tested.